MeetingMinutes.Client/python/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
av==17.0.0
braceexpand==0.1.7
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
ctranslate2==4.7.1
cuda-bindings
cuda-pathfinder==1.4.2
cycler==0.12.1
cytoolz==1.1.0
datasets==4.7.0
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
editdistance==0.8.1
einops==0.8.2
executing==2.2.1
faster-whisper==1.2.1
fiddle==0.3.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.73.0
graphviz==0.21
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
indic_numtowords==1.1.0
inflect==7.5.0
intervaltree==3.2.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.3
julius==0.2.7
kaldi-python-io==1.2.2
kaldialign==0.9.1
kiwisolver==1.5.0
lazy-loader==0.5
lhotse==1.32.2
libcst==1.8.6
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.3
lilcom==1.8.2
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.1.6
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
nemo-toolkit==2.7.0
networkx==3.6.1
numba==0.64.0
numba-cuda==0.15.1
numexpr==2.13.1
numpy==2.4.3
nv-one-logger-core==2.3.1
nv-one-logger-pytorch-lightning-integration==2.3.1
nv-one-logger-training-telemetry==2.3.1
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.24.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.8.0
overrides==7.7.0
packaging==24.2
pandas==3.0.1
parso==0.8.6
peft==0.18.1
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
portalocker==3.2.0
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
resampy==0.4.3
rich==14.3.3
ruamel.yaml==0.19.1
sacrebleu==2.6.0
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.54.0
six==1.17.0
smmap==5.0.3
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
SQLAlchemy==2.0.48
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
torch==2.10.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.10.0
torchcodec==0.10.0
torchmetrics==1.9.0
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.1
wcwidth==0.6.0
webdataset==1.0.2
Werkzeug==3.1.6
wget==3.2
whisper_normalizer==0.1.12
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
